Moved current location logic to its own class, and began writing a google map adapter
