+++ app/src/main/java/com/example/campusguide/Map.kt
@@ -1,0 +1,9 @@
+package com.example.campusguide
+
+import com.google.android.gms.maps.CameraUpdate
+import com.google.android.gms.maps.model.MarkerOptions
+
+interface Map {
+    fun addMarker(options: MarkerOptions)
+    fun animateCamera(update: CameraUpdate)
+}
