+++ app/src/main/java/com/example/campusguide/MapsActivity.kt
@@ -3,7 +3,6 @@ package com.example.campusguide
 import android.Manifest
 import android.content.Intent
 import android.content.pm.PackageManager
-import android.location.Location
 import android.os.Build
 import android.os.Bundle
 import android.view.View
@@ -16,14 +15,12 @@ import com.example.campusguide.directions.Route
 import com.example.campusguide.utils.BuildingHighlights
 import com.google.android.gms.common.ConnectionResult
 import com.google.android.gms.common.api.GoogleApiClient
-import com.google.android.gms.location.FusedLocationProviderClient
 import com.google.android.gms.location.LocationRequest
 import com.google.android.gms.location.LocationServices
 import com.google.android.gms.maps.CameraUpdateFactory
 import com.google.android.gms.maps.GoogleMap
 import com.google.android.gms.maps.OnMapReadyCallback
 import com.google.android.gms.maps.SupportMapFragment
-import com.google.android.gms.maps.model.BitmapDescriptorFactory
 import com.google.android.gms.maps.model.LatLng
 import com.google.android.gms.maps.model.MarkerOptions
 import com.google.android.libraries.places.api.Places
@@ -38,10 +35,10 @@ import kotlinx.android.synthetic.main.activity_maps.*
 class MapsActivity : AppCompatActivity(), OnMapReadyCallback,
     GoogleApiClient.ConnectionCallbacks, GoogleApiClient.OnConnectionFailedListener {
 
+    private val map = GoogleMapAdapter()
     private lateinit var mMap: GoogleMap
     private var mGoogleApiClient: GoogleApiClient? = null
     private lateinit var mLocationRequest: LocationRequest
-    private lateinit var fusedLocationClient: FusedLocationProviderClient
     private lateinit var route: Route
     private lateinit var buildingHighlights: BuildingHighlights
 
@@ -54,30 +51,12 @@ class MapsActivity : AppCompatActivity(), OnMapReadyCallback,
         mapFragment.getMapAsync(this)
 
         val currentLocationButton: FloatingActionButton = findViewById(R.id.currentLocationButton)
-        currentLocationButton.setOnClickListener {
-            //Check if location permission has been granted
-            if (ActivityCompat.checkSelfPermission(
-                    this,
-                    Manifest.permission.ACCESS_FINE_LOCATION
-                ) == PackageManager.PERMISSION_GRANTED
-            ) {
-                goToCurrentLocation()
-            } else {
-                //Request location permission
-                ActivityCompat.requestPermissions(
-                    this,
-                    arrayOf(Manifest.permission.ACCESS_FINE_LOCATION),
-                    Constants.LOCATION_PERMISSION_ACCESS_CODE
-                )
-            }
-        }
+        currentLocationButton.setOnClickListener(CenterLocationListener(this, map))
 
         ObjectBox.init(this.applicationContext)
       
         if (!Places.isInitialized())
             Places.initialize(applicationContext, getString(R.string.google_maps_key))
-
-        fusedLocationClient = LocationServices.getFusedLocationProviderClient(this)
     }
 
     /**
@@ -90,8 +69,9 @@ class MapsActivity : AppCompatActivity(), OnMapReadyCallback,
      * installed Google Play services and returned to the app.
      */
     override fun onMapReady(googleMap: GoogleMap) {
+        map.adapted = googleMap
         mMap = googleMap
-        mMap.uiSettings.isMyLocationButtonEnabled = false
+        googleMap.uiSettings.isMyLocationButtonEnabled = false
         route = Route(mMap, this)
 
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
@@ -152,33 +132,6 @@ class MapsActivity : AppCompatActivity(), OnMapReadyCallback,
         print("Connection has FAILED")
     }
 
-    /**
-     * Centers the map on the user's current location and places a marker.
-     */
-    private fun goToCurrentLocation() {
-        fusedLocationClient.lastLocation.addOnSuccessListener(this) { location ->
-            if(location != null) {
-                animateCurrentLocation(location)
-            }
-        }
-    }
-
-    private fun animateCurrentLocation(location: Location) {
-        val currentLatLng = LatLng(location.latitude, location.longitude)
-        mMap.addMarker(
-            MarkerOptions()
-                .position(currentLatLng)
-                .title("You are here.")
-                .icon(BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_AZURE))
-        )
-        mMap.animateCamera(
-            CameraUpdateFactory.newLatLngZoom(
-                currentLatLng,
-                Constants.ZOOM_STREET_LVL
-            )
-        )
-    }
-
     override fun onRequestPermissionsResult(
         requestCode: Int,
         permissions: Array<out String>,
@@ -188,7 +141,7 @@ class MapsActivity : AppCompatActivity(), OnMapReadyCallback,
         when (requestCode) {
             Constants.LOCATION_PERMISSION_ACCESS_CODE -> {
                 if ((grantResults.isNotEmpty() && grantResults[0] == PackageManager.PERMISSION_GRANTED)) {
-                    goToCurrentLocation()
+                    //goToCurrentLocation()
                 }
                 return
             }
@@ -244,32 +197,9 @@ class MapsActivity : AppCompatActivity(), OnMapReadyCallback,
      * Methods for setting button listeners all at once.
      */
     private fun setButtonListeners() {
-        setCurrentLocationListener()
         setNavButtonListener()
         val switchCampusToggle: ToggleButton = findViewById(R.id.switchCampusButton)
         SwitchCampus(switchCampusToggle, mMap, campus_name)
-    }
-
-    private fun setCurrentLocationListener() {
-        val currentLocationButton: FloatingActionButton =
-            findViewById(R.id.currentLocationButton)
-        currentLocationButton.setOnClickListener {
-            //Check if location permission has been granted
-            if (ActivityCompat.checkSelfPermission(
-                    this,
-                    Manifest.permission.ACCESS_FINE_LOCATION
-                ) == PackageManager.PERMISSION_GRANTED
-            ) {
-                goToCurrentLocation()
-            } else {
-                //Request location permission
-                ActivityCompat.requestPermissions(
-                    this,
-                    arrayOf(Manifest.permission.ACCESS_FINE_LOCATION),
-                    Constants.LOCATION_PERMISSION_ACCESS_CODE
-                )
-            }
-        }
     }
 
     fun onOpenMenu(view: View) {
